Update ListResponse#getTotalResults() to return int instead of long
This aligns the return type of this method with the class's constructor.

Fixes issue #117.

+++ scim2-sdk-common/src/main/java/com/unboundid/scim2/common/messages/ListResponse.java
@@ -150,7 +150,7 @@ public final class ListResponse<T> extends BaseScimResource
    * @return The total number of results returned by the list or query
    * operation.
    */
-  public long getTotalResults()
+  public int getTotalResults()
   {
     return totalResults;
   }
